pyproject.toml
[project]
name = "vibe-blog"
version = "0.1.0"
description = "VibeBlog - 技术科普绘本生成器"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web 框架
    "flask-cors>=4.0.0",
    "flask>=3.0.0",
    "gunicorn>=21.0.0",
    # 控制台彩色日志
    "rich>=13.7.0",
    # LLM集成
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-google-genai>=1.0.0",
    # LangGraph 工作流引擎 (博客生成)
    "langgraph>=1.0.0",
    # 模板引擎 (Prompt 管理)
    "jinja2>=3.1.0",
    # HTTP 请求
    "requests>=2.31.0",
    # 环境变量
    "python-dotenv>=1.0.1",
    # 数据验证
    "pydantic>=2.0.0",
    # 图片处理
    "pillow>=10.0.0",
    # Word 导出
    "python-docx>=1.1.0",
    # 阿里云 OSS
    "oss2>=2.18.0",
    # 中文分词（用于可读性分析）
    # 注：readability-cn 依赖太重（需要 LTP/torch），使用自定义分析器代替
    "jieba>=0.42.1",
    # 多平台发布
    "playwright>=1.40.0",
    "pyyaml>=6.0",
    # Langfuse 追踪
    "langfuse>=3.0.0",
    "opentelemetry-instrumentation-threading>=0.48b0",
    # 任务排队 & 定时调度
    "aiosqlite>=0.19.0",
    "croniter>=6.0.0",
    # 保留兼容，渐进迁移后可移除
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
test = [
    "pytest-cov>=4.1.0",
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.0",
    "pytest-flask>=1.3.0",
    "pytest-xdist>=3.5.0",
]


[tool.pytest.ini_options]
pythonpath = [".", "vibe-blog/backend"]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

# 命令行选项（addopts）
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=services",
    "--cov=routes",
    "--cov=utils",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=json",
    "--cov-report=lcov",
]

# 标记（markers）
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may use database, file system)",
    "api: API endpoint tests (Flask test client)",
    "llm: Tests that require LLM API calls (expensive, skip in CI)",
    "slow: Slow tests (skip by default)",
]

# asyncio 配置
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# 警告过滤
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::ResourceWarning",
]

# 覆盖率阈值（属于 coverage 工具，不在 pytest.ini_options 内）
[tool.coverage.report]
fail_under = 55
